[Style] #44 - HeaderView 분리
